Fix bigint support test on Lua versions without the floor division operator.
Some distros have Love linked to LuaJIT, which is currently equivalent to
Lua 5.1 and thus does not support the `//` floor division operator introduced
in Lua 5.4. In this case the code that tests for bigint support
fails to compile, so `loadString` returns `nil`, which cannot be invoked
as a function.

This commit adds a `nil` check for the result of `loadString` and treats
such runtimes as not supporting bigints.